+++ DataAccess/Repository/IRepository/IArticleRepository.cs
@@ -10,7 +10,7 @@ namespace DataAccess.Repository.IRepository
     public interface IArticleRepository
     {
         Task<List<Article>> GetAll();
-        Task<Article?> FindById(int id);
+        Task<Article?> FindById(long id);
         Task<Article?> FindByTitle(string Title);
         Task<Article?> FindBySlug(string slug);
         Task Add(Article article);
